Added password hashing and salting. Using Argon2 algorithm with the dependency "de.mkammerer:argon2-jvm".
A better option would be to use the argon2-jvm-nolibs on a Linux server (which would require the separate installation of argon2 libs), but at the moment for testing purposes on local Windows machines I've implemented the version with built-in argon2.

Using argon2 package would be a better solution because it is maintained more than de.mkammerer Java library.